mon/PGMap: fix digest summary of PGs by state
This fixes the PG state summary seen in 'ceph -s' to reflect PG state bits
beyond 32 bits.

Fixes: https://tracker.ceph.com/issues/44052
(cherry picked from commit 1453154c252e617efd01fce59c71a5dcc3a73c23)

Conflicts:
	src/mon/PGMap.cc
- nautilus has "multimap" where master has "std::multimap"

+++ src/mon/PGMap.cc
@@ -214,7 +214,7 @@ void PGMapDigest::print_summary(ceph::Formatter *f, ostream *out) const
     f->open_array_section("pgs_by_state");
 
   // list is descending numeric order (by count)
-  std::multimap<int,int> state_by_count;  // count -> state
+  std::multimap<int,uint64_t> state_by_count;  // count -> state
   for (auto p = num_pg_by_state.begin();
        p != num_pg_by_state.end();
        ++p) {
